+++ confignode/src/main/java/org/apache/iotdb/confignode/service/ConfigNode.java
@@ -20,8 +20,8 @@ package org.apache.iotdb.confignode.service;
 
 import org.apache.iotdb.common.rpc.thrift.TConfigNodeLocation;
 import org.apache.iotdb.common.rpc.thrift.TEndPoint;
+import org.apache.iotdb.commons.conf.CommonDescriptor;
 import org.apache.iotdb.commons.exception.BadNodeUrlException;
-import org.apache.iotdb.commons.exception.ConfigurationException;
 import org.apache.iotdb.commons.exception.StartupException;
 import org.apache.iotdb.commons.service.JMXService;
 import org.apache.iotdb.commons.service.RegisterManager;
@@ -29,23 +29,31 @@ import org.apache.iotdb.commons.udf.service.UDFClassLoaderManager;
 import org.apache.iotdb.commons.udf.service.UDFExecutableManager;
 import org.apache.iotdb.commons.udf.service.UDFRegistrationService;
 import org.apache.iotdb.commons.utils.NodeUrlUtils;
+import org.apache.iotdb.confignode.client.SyncConfigNodeClientPool;
 import org.apache.iotdb.confignode.conf.ConfigNodeConfig;
 import org.apache.iotdb.confignode.conf.ConfigNodeConstant;
 import org.apache.iotdb.confignode.conf.ConfigNodeDescriptor;
 import org.apache.iotdb.confignode.conf.ConfigNodeRemoveCheck;
-import org.apache.iotdb.confignode.conf.ConfigNodeStartupCheck;
+import org.apache.iotdb.confignode.conf.SystemPropertiesUtils;
 import org.apache.iotdb.confignode.manager.ConfigManager;
+import org.apache.iotdb.confignode.rpc.thrift.TConfigNodeRegisterReq;
+import org.apache.iotdb.confignode.rpc.thrift.TConfigNodeRegisterResp;
 import org.apache.iotdb.confignode.service.thrift.ConfigNodeRPCService;
 import org.apache.iotdb.confignode.service.thrift.ConfigNodeRPCServiceProcessor;
 import org.apache.iotdb.db.service.metrics.MetricsService;
+import org.apache.iotdb.rpc.TSStatusCode;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
+import java.util.concurrent.TimeUnit;
 
 public class ConfigNode implements ConfigNodeMBean {
+
   private static final Logger LOGGER = LoggerFactory.getLogger(ConfigNode.class);
+
+  private static final ConfigNodeConfig conf = ConfigNodeDescriptor.getInstance().getConf();
 
   private final String mbeanName =
       String.format(
@@ -53,9 +61,6 @@ public class ConfigNode implements ConfigNodeMBean {
           ConfigNodeConstant.CONFIGNODE_PACKAGE, ConfigNodeConstant.JMX_TYPE, "ConfigNode");
 
   private final RegisterManager registerManager = new RegisterManager();
-
-  private ConfigNodeRPCService configNodeRPCService;
-  private ConfigNodeRPCServiceProcessor configNodeRPCServiceProcessor;
 
   private ConfigManager configManager;
 
@@ -68,29 +73,52 @@ public class ConfigNode implements ConfigNodeMBean {
   }
 
   public void active() {
+    LOGGER.info("Activating {}...", ConfigNodeConstant.GLOBAL_NAME);
+
     try {
-      // Set up services
-      setUpServices();
-      // Do ConfigNode startup checks
-      ConfigNodeStartupCheck.getInstance().startUpCheck();
       // Init ConfigManager
       initConfigManager();
-    } catch (StartupException | IOException | ConfigurationException e) {
+      // Set up services
+      setUpServices();
+
+      // Check for initial startup or restart
+      if (SystemPropertiesUtils.isRestarted()) {
+        // Restart process finished
+        LOGGER.info(
+            "{} has successfully started and joined the cluster.", ConfigNodeConstant.GLOBAL_NAME);
+      } else if (ConfigNodeDescriptor.getInstance().isSeedConfigNode()) {
+        SystemPropertiesUtils.storeSystemParameters();
+        // Seed-ConfigNode should apply itself when first start
+        configManager
+            .getNodeManager()
+            .applyConfigNode(
+                new TConfigNodeLocation(
+                    0,
+                    new TEndPoint(conf.getRpcAddress(), conf.getRpcPort()),
+                    new TEndPoint(conf.getRpcAddress(), conf.getConsensusPort())));
+        // The initial startup of Seed-ConfigNode finished
+        LOGGER.info(
+            "{} has successfully started and joined the cluster.", ConfigNodeConstant.GLOBAL_NAME);
+      } else {
+        registerConfigNode();
+        // The initial startup of Non-Seed-ConfigNode is not yet finished,
+        // We should wait for leader's scheduling
+        LOGGER.info(
+            "{} has registered successfully. Waiting for the leader's scheduling to join the cluster.",
+            ConfigNodeConstant.GLOBAL_NAME);
+      }
+
+    } catch (StartupException | IOException e) {
       LOGGER.error("Meet error while starting up.", e);
       try {
-        deactivate();
+        stop();
       } catch (IOException e2) {
         LOGGER.error("Meet error when stop ConfigNode!", e);
       }
-      return;
     }
-
-    LOGGER.info(
-        "{} has successfully started and joined the cluster.", ConfigNodeConstant.GLOBAL_NAME);
   }
 
   private void initConfigManager() {
-    // Init ConfigManager
     try {
       configManager = new ConfigManager();
     } catch (IOException e) {
@@ -102,42 +130,79 @@ public class ConfigNode implements ConfigNodeMBean {
       }
       System.exit(-1);
     }
+    configManager.addMetrics();
 
-    // Init RPC service
-    configNodeRPCService = new ConfigNodeRPCService();
-    configNodeRPCServiceProcessor = new ConfigNodeRPCServiceProcessor(configManager);
+    LOGGER.info("Successfully initialize ConfigManager.");
   }
 
   private void setUpServices() throws StartupException, IOException {
-    LOGGER.info("Setting up services of {}...", ConfigNodeConstant.GLOBAL_NAME);
-
+    // Setup JMXService
     registerManager.register(new JMXService());
     JMXService.registerMBean(this, mbeanName);
 
-    registerManager.register(MetricsService.getInstance());
-    configManager.addMetrics();
-    registerUdfServices();
+    // Setup UDFService
+    registerManager.register(
+        UDFExecutableManager.setupAndGetInstance(conf.getTemporaryLibDir(), conf.getUdfLibDir()));
+    registerManager.register(UDFClassLoaderManager.setupAndGetInstance(conf.getUdfLibDir()));
+    registerManager.register(UDFRegistrationService.setupAndGetInstance(conf.getSystemUdfDir()));
 
+    // Setup MetricsService
+    registerManager.register(MetricsService.getInstance());
+    MetricsService.getInstance().startAllReporter();
+
+    // Setup RPCService
+    ConfigNodeRPCService configNodeRPCService = new ConfigNodeRPCService();
+    ConfigNodeRPCServiceProcessor configNodeRPCServiceProcessor =
+        new ConfigNodeRPCServiceProcessor(configManager);
     configNodeRPCService.initSyncedServiceImpl(configNodeRPCServiceProcessor);
     registerManager.register(configNodeRPCService);
-    LOGGER.info("Init rpc server success");
 
-    // start reporter
-    MetricsService.getInstance().startAllReporter();
+    LOGGER.info("Successfully setup internal services.");
   }
 
-  private void registerUdfServices() throws StartupException {
-    final ConfigNodeConfig configNodeConfig = ConfigNodeDescriptor.getInstance().getConf();
-    registerManager.register(
-      UDFExecutableManager.setupAndGetInstance(
-        configNodeConfig.getTemporaryLibDir(), configNodeConfig.getUdfLibDir()));
-    registerManager.register(
-      UDFClassLoaderManager.setupAndGetInstance(configNodeConfig.getUdfLibDir()));
-    registerManager.register(
-      UDFRegistrationService.setupAndGetInstance(configNodeConfig.getSystemUdfDir()));
+  /** Register Non-seed ConfigNode when first startup */
+  private void registerConfigNode() throws StartupException {
+    TConfigNodeRegisterReq req =
+        new TConfigNodeRegisterReq(
+            new TConfigNodeLocation(
+                -1,
+                new TEndPoint(conf.getRpcAddress(), conf.getRpcPort()),
+                new TEndPoint(conf.getRpcAddress(), conf.getConsensusPort())),
+            conf.getDataRegionConsensusProtocolClass(),
+            conf.getSchemaRegionConsensusProtocolClass(),
+            conf.getSeriesPartitionSlotNum(),
+            conf.getSeriesPartitionExecutorClass(),
+            CommonDescriptor.getInstance().getConfig().getDefaultTTL(),
+            conf.getTimePartitionInterval(),
+            conf.getSchemaReplicationFactor(),
+            conf.getSchemaRegionPerDataNode(),
+            conf.getDataReplicationFactor(),
+            conf.getDataRegionPerProcessor());
+
+    TEndPoint targetConfigNode = conf.getTargetConfigNode();
+    while (true) {
+      TConfigNodeRegisterResp resp =
+          SyncConfigNodeClientPool.getInstance().registerConfigNode(targetConfigNode, req);
+      if (resp.getStatus().getCode() == TSStatusCode.SUCCESS_STATUS.getStatusCode()) {
+        conf.setPartitionRegionId(resp.getPartitionRegionId().getId());
+        break;
+      } else if (resp.getStatus().getCode() == TSStatusCode.NEED_REDIRECTION.getStatusCode()) {
+        targetConfigNode = resp.getStatus().getRedirectNode();
+        LOGGER.info("ConfigNode need redirect to  {}.", targetConfigNode);
+      } else if (resp.getStatus().getCode() == TSStatusCode.ERROR_GLOBAL_CONFIG.getStatusCode()) {
+        LOGGER.error("Configuration may not be consistent, {}", req);
+        throw new StartupException("Configuration are not consistent!");
+      }
+
+      try {
+        TimeUnit.MILLISECONDS.sleep(1000);
+      } catch (InterruptedException e) {
+        throw new StartupException("Register ConfigNode failed!");
+      }
+    }
   }
 
-  public void deactivate() throws IOException {
+  public void stop() throws IOException {
     LOGGER.info("Deactivating {}...", ConfigNodeConstant.GLOBAL_NAME);
     registerManager.deregisterAll();
     JMXService.deregisterMBean(mbeanName);
@@ -145,10 +210,6 @@ public class ConfigNode implements ConfigNodeMBean {
       configManager.close();
     }
     LOGGER.info("{} is deactivated.", ConfigNodeConstant.GLOBAL_NAME);
-  }
-
-  public void stop() throws IOException {
-    deactivate();
   }
 
   public void doRemoveNode(String[] args) throws IOException {
